+++ http.cpp
@@ -38,8 +38,8 @@ bool tokenize( char** buffer, const char* bufferEnd, char** token ){
 
 
 // look for a line like: GET /?QUERY=2036&SPORT=33344&PORT_EQ_2132 HTTP/1.1
-int Request::parse(Sock& conn, vector<string>& filters, unsigned short& port){
-    long int query = 0;
+int Request::parse(Sock& conn, vector<string>& filters, uint16_t& port){
+    uint32_t query = 0;
     char buff[2048];
     for(int i=0; i< 10; ++i) { // drop if the other side is slow to send request
         if(! conn.readLine(buff, sizeof(buff) ) ) {
@@ -68,13 +68,13 @@ int Request::parse(Sock& conn, vector<string>& filters, unsigned short& port){
 }
 
 
-void turnBitsOff(int& mask, int bits){
+void turnBitsOff(uint32_t& mask, uint32_t bits){
     mask = mask & (0xFFFFFFFF^bits); // TODO: return a new mask instead of taking a reference???
 }
 
-int Response::write(Sock& conn, int select, vector<string>& filters, LocalData& ldata, RemoteData& rdata, BWLists& bwlists ){
+int Response::write(Sock& conn, uint32_t select, vector<string>& filters, LocalData& ldata, RemoteData& rdata, BWLists& bwlists ){
     static const string header =  "HTTP/1.1 200 OK\n\n";
-    size_t bytes = conn.write(header.c_str(), header.size() ); // no need to sign the header
+    int bytes = conn.write(header.c_str(), header.size() ); // no need to sign the header
 
     // If you don't want to share some requested fieds, turn off the bits.
     turnBitsOff(select, SELECTION::WLKEY);  // do not share your pub key white list (friends)
@@ -84,11 +84,11 @@ int Response::write(Sock& conn, int select, vector<string>& filters, LocalData& 
     Writer* writer = sign ? &signatureWriter : &dumbWriter;
     writer->init(conn);
 
-    unsigned int netSelect = htonl(select);
+    uint32_t netSelect = htonl(select);
     bytes+= writer->write((char*) &netSelect, sizeof(netSelect));
 
     if( select & SELECTION::MYIP ){
-        unsigned long remoteIP = conn.getIP();                    // remote IP the way I see it
+        uint32_t remoteIP = conn.getIP();                    // remote IP the way I see it
         bytes+=writer->write((char*)&remoteIP, sizeof(remoteIP)); // helps other server find it's NATed IPs
     }
 
@@ -104,7 +104,7 @@ int Response::write(Sock& conn, int select, vector<string>& filters, LocalData& 
 }
 
 
-void Response::writeDebug(Sock& conn, int select, std::vector<std::string>& filters){
+void Response::writeDebug(Sock& conn, uint32_t select, std::vector<std::string>& filters){
     stringstream ss;
     ss << "HTTP/1.1 200 OK\n";
     ss << "Server: n3+1\n\n"; // "\n" separates headers from html
